Moves components around, deletes rendering until I figure out a better way to segment the project, updates readme.
