+++ Core/Components/Anchor.cs
@@ -2,7 +2,7 @@ using Unity.Entities;
 using Unity.Mathematics;
 using UnityEngine;
 
-namespace UGUIDots {
+namespace UGUIDots.Transforms {
 
     /// <summary>
     /// Flags which state how a transform is anchored relative to the resolution of the screen.
