+++ Core/Systems/Conversions/RectTransformConversionSystem.cs
@@ -1,3 +1,4 @@
+using UGUIDots.Transforms;
 using UnityEngine;
 
 namespace UGUIDots.Conversions.Systems {
