--- Rendering.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f84ef99b89add424293244f8798ce625
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
