--- Rendering/UGUIDots.Rendering.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "UGUIDots.Rendering"
-}
