--- Rendering/UGUIDots.Rendering.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5d5a385f587b04cbe9c54e50d1274441
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
